Added Set NBT Command for command blocks
I wanted to edit custom named portal nbt data, but you normally have to look at the portal, so I mixed the set portal nbt data command and one of the command block commands to create this!

Also my editor thought it was a good idea to auto format all of the code in the file, so yeah, its formatted now.